Updating tasks to allow for actually using them
The tasks cog was broken in a few ways
  - the tasks loop ate any exceptions and then hammered them repeatedly
    until it could get success, willing to be you see increased api hits
yesterday.
  - Missing imports
  - bot vs self.bot issues

Updated/Fixed all of these and added better logging surrounding the
running of the tasks. It now will send a "Started" message and a
"Completed" message. In addition the entire task is wrapped in a
try/except to always return okay.

I also updated the league one to ignore failed inserts, so even if a
team does not exist it can still insert all of the matches not impacted
by that issue.